Add files via upload
This is a basic web bookmark folder structure that follows the Lambda - Web Dev course schedule as of 8/4/20.
I made it for the reason I cant think of a better way to organize the many bookmarks one collects during the course of attending Lambda School.
There may be a much better and simpler way and I'd be glad to hear any and all ideas.
It has a few links added which I considered necessary purely as examples.  
The import method is browser specific so personal research may be needed to add them to your browser.